可視範囲のspots再取得ボタンの機能実装
