+++ lib/views/widgets/search_button.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_map_app/providers/location_provider.dart';
+import 'package:flutter_map_app/providers/markers_provider.dart';
+import 'package:flutter_map_app/theme.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class SearchButton extends ConsumerWidget {
+  const SearchButton({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final currentBounds = ref.watch(mapBoundsProvider);
+    return Container(
+      height: 60,
+      margin: const EdgeInsets.only(top: 50, left: 15, right: 15),
+      child: ElevatedButton(
+        onPressed: () {
+          // 再検索時に充電スポットを取得してマーカーを更新
+          if (currentBounds != null) {
+            ref
+                .read(markerProvider.notifier)
+                .searchSpots(ref: ref, bounds: currentBounds);
+          }
+        },
+        style: ElevatedButton.styleFrom(
+          backgroundColor: ThemeColor.lightGreen,
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(34),
+          ),
+        ),
+        child: const Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Text(
+              'このエリアでスポット検索',
+              style: TextStyle(
+                color: ThemeColor.green,
+              ),
+            ),
+            Icon(
+              Icons.search,
+              color: ThemeColor.green,
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
